✨[Feat] mapRecordUpdateToUpdateBoardInput(recordUpdateInputValue->recordUpdateArgs) 함수 구현
